DynamixelIO --> DynamixelProtocolHandler
Cleanup API and naming.
