+++ src/device/mx.rs
@@ -50,11 +50,11 @@ reg_read_write!(goal_acceleration, 73, u8);
 ///
 /// reg_read_only!(present_position_speed_load, 36, (i16, u16, u16))
 pub fn sync_read_present_position_speed_load(
-    io: &DynamixelSerialIO,
+    dph: &DynamixelProtocolHandler,
     serial_port: &mut dyn serialport::SerialPort,
     ids: &[u8],
 ) -> Result<Vec<(i16, u16, u16)>> {
-    let val = io.sync_read(serial_port, ids, 36, 2 + 2 + 2)?;
+    let val = dph.sync_read(serial_port, ids, 36, 2 + 2 + 2)?;
     let val = val
         .iter()
         .map(|v| {
